multi: move rfq channel helpers to rfq manager
We now move some of the rfq channel helpers that previously lived in
rpcserver to the rfq package. We also use the new channel lister we
introduced previously.
